:adhesive_bandage: MemberService Interface 생성
